nano_malloc_usable_size and nano_realloc: fix long->int truncation.
As the size/offset is a long it should not be truncated to int in
nano_malloc_usable_size.

+++ newlib/libc/stdlib/nano-mallocr.c
@@ -525,7 +525,7 @@ void nano_malloc_stats(RONEARG)
 malloc_size_t nano_malloc_usable_size(RARG void * ptr)
 {
     chunk * c = (chunk *)((char *)ptr - CHUNK_OFFSET);
-    int size_or_offset = c->size;
+    long size_or_offset = c->size;
 
     if (size_or_offset < 0)
     {
